+++ KubeArmor/audit/auditLogger.go
@@ -28,21 +28,26 @@ type AuditLogger struct {
 
 	// container id -> cotnainer
 	Containers     *map[string]tp.Container
-	ContainersLock **sync.Mutex
+	ContainersLock **sync.RWMutex
 
 	// container id -> (host) pid
 	ActivePidMap     *map[string]tp.PidMap
 	ActiveHostPidMap *map[string]tp.PidMap
+	ActivePidMapLock **sync.RWMutex
 
-	// pid map lock
-	ActivePidMapLock **sync.Mutex
+	// host pid -> host pid
+	ActiveHostMap     *map[uint32]tp.PidMap
+	ActiveHostMapLock **sync.RWMutex
 
 	// GKE
 	IsCOS bool
 }
 
 // NewAuditLogger Function
-func NewAuditLogger(feeder *fd.Feeder, HomeDir string, containers *map[string]tp.Container, containersLock **sync.Mutex, activePidMap *map[string]tp.PidMap, activeHostPidMap *map[string]tp.PidMap, activePidMapLock **sync.Mutex) *AuditLogger {
+func NewAuditLogger(feeder *fd.Feeder,
+	containers *map[string]tp.Container, containersLock **sync.RWMutex,
+	activePidMap *map[string]tp.PidMap, activeHostPidMap *map[string]tp.PidMap, activePidMapLock **sync.RWMutex,
+	activeHostMap *map[uint32]tp.PidMap, activeHostMapLock **sync.RWMutex) *AuditLogger {
 	adt := new(AuditLogger)
 
 	adt.HostName = kl.GetHostName()
@@ -55,6 +60,9 @@ func NewAuditLogger(feeder *fd.Feeder, HomeDir string, containers *map[string]tp
 	adt.ActivePidMap = activePidMap
 	adt.ActiveHostPidMap = activeHostPidMap
 	adt.ActivePidMapLock = activePidMapLock
+
+	adt.ActiveHostMap = activeHostMap
+	adt.ActiveHostMapLock = activeHostMapLock
 
 	adt.IsCOS = false
 
@@ -141,164 +149,6 @@ func (adt *AuditLogger) DestroyAuditLogger() error {
 // == Auditd == //
 // ============ //
 
-// GetProcessInfoFromHostPid Function
-func (adt *AuditLogger) GetProcessInfoFromHostPid(log tp.Log, hostPid int32) tp.Log {
-	ActiveHostPidMap := *(adt.ActiveHostPidMap)
-
-	ActivePidMapLock := *(adt.ActivePidMapLock)
-	ActivePidMapLock.Lock()
-	defer ActivePidMapLock.Unlock()
-
-	for id, pidMap := range ActiveHostPidMap {
-		for pid, node := range pidMap {
-			if hostPid == int32(pid) {
-				log.ContainerID = id
-
-				log.PPID = int32(node.PPID)
-				log.PID = int32(node.PID)
-				log.UID = int32(node.UID)
-
-				break
-			}
-		}
-
-		if log.ContainerID != "" {
-			break
-		}
-	}
-
-	if log.ContainerID == "" {
-		for id, pidMap := range ActiveHostPidMap {
-			for pid, node := range pidMap {
-				if hostPid == int32(pid) {
-					log.ContainerID = id
-
-					log.PPID = int32(node.PPID)
-					log.PID = int32(node.PID)
-					log.UID = int32(node.UID)
-
-					break
-				}
-			}
-
-			if log.ContainerID != "" {
-				break
-			}
-		}
-	}
-
-	if log.PID == 0 {
-		log.PPID = -1
-		log.PID = -1
-		log.UID = -1
-	}
-
-	return log
-}
-
-// GetContainerInfoFromContainerID Function
-func (adt *AuditLogger) GetContainerInfoFromContainerID(log tp.Log, profileName string) tp.Log {
-	Containers := *(adt.Containers)
-	ContainersLock := *(adt.ContainersLock)
-
-	ContainersLock.Lock()
-	defer ContainersLock.Unlock()
-
-	if log.ContainerID != "" {
-		if val, ok := Containers[log.ContainerID]; ok {
-			log.NamespaceName = val.NamespaceName
-			log.PodName = val.ContainerGroupName
-			log.ContainerName = val.ContainerName
-		}
-	} else {
-		for _, container := range Containers {
-			if strings.HasPrefix(profileName, container.AppArmorProfile) {
-				log.NamespaceName = container.NamespaceName
-				log.PodName = container.ContainerGroupName
-				log.ContainerID = container.ContainerID
-				log.ContainerName = container.ContainerName
-				break
-			}
-		}
-	}
-
-	return log
-}
-
-// GetExecPath Function
-func (adt *AuditLogger) GetExecPath(containerID string, pid uint32) string {
-	ActivePidMap := *(adt.ActivePidMap)
-
-	ActivePidMapLock := *(adt.ActivePidMapLock)
-	ActivePidMapLock.Lock()
-	defer ActivePidMapLock.Unlock()
-
-	if pidMap, ok := ActivePidMap[containerID]; ok {
-		if node, ok := pidMap[pid]; ok {
-			if node.PID == pid {
-				return node.ExecPath
-			}
-		}
-	}
-
-	return ""
-}
-
-// UpdateSourceAndResource Function
-func (adt *AuditLogger) UpdateSourceAndResource(log tp.Log, source, resource string) tp.Log {
-	if log.Operation == "Process" {
-		log.Source = adt.GetExecPath(log.ContainerID, uint32(log.PPID))
-		if log.Source == "" {
-			log.Source = source
-		}
-
-		log.Resource = adt.GetExecPath(log.ContainerID, uint32(log.PID))
-		if log.Resource == "" {
-			log.Resource = resource
-		} else if !strings.HasPrefix(log.Resource, resource) {
-			log.Resource = resource
-		}
-	} else { // File
-		log.Source = adt.GetExecPath(log.ContainerID, uint32(log.PID))
-		if log.Source == "" {
-			log.Source = source
-		}
-
-		log.Resource = resource
-	}
-
-	return log
-}
-
-// GetHostProcessInfoFromHostPid Function
-func (adt *AuditLogger) GetHostProcessInfoFromHostPid(log tp.Log, hostPid int32) tp.Log {
-	ActiveHostPidMap := *(adt.ActiveHostPidMap)
-
-	ActivePidMapLock := *(adt.ActivePidMapLock)
-	ActivePidMapLock.Lock()
-	defer ActivePidMapLock.Unlock()
-
-	for _, pidMap := range ActiveHostPidMap {
-		for pid, node := range pidMap {
-			if hostPid == int32(pid) {
-				log.PPID = int32(node.PPID)
-				log.PID = int32(node.PID)
-				log.UID = int32(node.UID)
-
-				break
-			}
-		}
-	}
-
-	if log.PID == 0 {
-		log.PPID = -1
-		log.PID = hostPid
-		log.UID = -1
-	}
-
-	return log
-}
-
 // GenerateAuditLog Function
 func (adt *AuditLogger) GenerateAuditLog(hostPid int32, profileName, source, operation, resource, action, data string) {
 	log := tp.Log{}
@@ -309,14 +159,14 @@ func (adt *AuditLogger) GenerateAuditLog(hostPid int32, profileName, source, ope
 	log.HostPID = hostPid
 	log.Operation = operation
 
-	if profileName == "kubearmor.host" {
-		log = adt.GetHostProcessInfoFromHostPid(log, hostPid) // PPID, PID, UID
-	} else {
-		log = adt.GetProcessInfoFromHostPid(log, hostPid)           // ContainerID, PPID, PID, UID
+	if profileName == "kubearmor.host" { // host
+		log = adt.GetHostProcessInfoFromHostPid(log, uint32(hostPid)) // PPID, PID, UID
+		log = adt.UpdateHostSourceAndResource(log, source, resource)  // Source, Resource
+	} else { // containers
+		log = adt.GetProcessInfoFromHostPid(log, uint32(hostPid))   // ContainerID, PPID, PID, UID
 		log = adt.GetContainerInfoFromContainerID(log, profileName) // NamespaceName, PodName, ContainerName
+		log = adt.UpdateSourceAndResource(log, source, resource)    // Source, Resource
 	}
-
-	log = adt.UpdateSourceAndResource(log, source, resource) // Source, Resource
 
 	log.Data = data
 
@@ -350,7 +200,7 @@ func (adt *AuditLogger) MonitorAuditLogs() {
 
 		if !strings.Contains(line, "AVC") {
 			continue
-		} else if !strings.Contains(line, "DENIED") && !strings.Contains(line, "AUDIT") {
+		} else if !strings.Contains(line, "DENIED") { // && !strings.Contains(line, "AUDIT") {
 			continue
 		} else if !strings.Contains(line, "exec") && !strings.Contains(line, "open") {
 			continue
